+++ src/App.js
@@ -1,29 +1,14 @@
 import React from 'react';
-const todoList = [
-  {
-    id: 1,
-    title: "Reed the book",
-  }, {
-    id: 2,
-    title: "Watch the videos",
-  }, {
-    id: 3,
-    title: "Complete assignment"
-  } ];
+import TodoList from './TodoList';
+import AddTodoForm from './AddTodoForm';
+
 
 function App() {
   return (
     <div>
       <h1>Todo List</h1>
-        <ul>
-          hint: {todoList.map (function (item) {
-          return <li key = {item.id}>
-                    <span>{item.id}</span>
-                    <span>{item.title}</span>
-                 </li>;
-                
-        })}
-        </ul>
+        <AddTodoForm/>
+        <TodoList/>
     </div>
     );
 }
